refactor: reorganize styles and remove redundant LoadingPart component
Moved CSS files to a dedicated `styles` directory for better organization. Removed the `LoadingPart` component and consolidated its functionality into `LoadingScreen` to reduce redundancy and improve maintainability. Updated import paths across the application to reflect the new structure.
