+++ src/components/ClanMaps/ResourceMapNoLog.tsx
@@ -8,7 +8,7 @@ import ModalMessage from "../ModalMessage";
 import MapLayer from "./MapLayer";
 import ResourcesInMapList from "./ResourcesInMapList";
 import CreateResourceTab from "./CreateResourceTab";
-import "../../css/map-sidebar.css";
+import "../../styles/map-sidebar.css";
 import { useLocation, useParams } from "react-router";
 import type { Marker } from "../../types/dto/marker";
 import type { ResourceInfo } from "../../types/dto/resources";
