+++ src/styles/loading-screen.css
@@ -1,19 +1,3 @@
-body {
-  background-color: #222;
-}
-.preloader {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-}
-.preloader-part {
-  width: 150px;
-  height: 150px;
-  margin: auto;
-}
-
 .loader {
   display: block;
   position: relative;
